pyproject.toml
[tool.poetry]
name = "django-playground"
version = "0.1.0"
description = ""
authors = ["Kosuke Saigusa <saigusa758cloudy@gmail.com>"]
readme = "README.md"
packages = [{include = "django_playground"}]

[tool.poetry.dependencies]
python = "^3.9"
Django = "3.2.15"
djangorestframework = "^3.14.0"


[tool.poetry.group.dev.dependencies]
black = "^22.10.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
